Fix sample_rate type mismatches for LocalAlbum and inserts
Changed LocalAlbum.sample_rate from u32 to f64 to match database REAL type.
Fixed database queries to read sample_rate as f64 instead of u32.
Removed incorrect i64 casts when inserting sample_rate values.

This fixes the 'Invalid column type Real' error that prevented the app
from loading after the database migration.

+++ src-tauri/src/library/models.rs
@@ -128,7 +128,7 @@ pub struct LocalAlbum {
     pub total_duration_secs: u64,
     pub format: AudioFormat,
     pub bit_depth: Option<u32>,
-    pub sample_rate: u32,
+    pub sample_rate: f64,  // Changed from u32 to f64 for decimal precision
     pub directory_path: String,
 }
